+++ CHANGES.rst
@@ -4,6 +4,7 @@
 - #30: Added :attr:`pyudev.Device.sys_number`
 - #29: Added :meth:`pyudev.Device.from_device_number` and
   :attr:`pyudev.Device.device_number`
+- Officially support PyPy now
 
 
 0.10 (Apr 20, 2011)